refactor: _components 폴더이동 및 코드정리
